requirements.txt
click==8.0.3
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap4==4.0.2
Flask-WTF==1.0.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.0
